[level 0] Title: 중앙값 구하기, Time: 0.32 ms, Memory: 74.6 MB -BaekjoonHub

+++ 프로그래머스/lv0/120811. 중앙값 구하기/중앙값 구하기.java
@@ -2,6 +2,6 @@ import java.util.Arrays;
 class Solution {
     public int solution(int[] array) {
         Arrays.sort(array);
-        return array[(int) array.length / 2];
+        return array[array.length / 2];
     }
 }
